fix(pwa): Stop calling supported browsers unsupported
Changes:
- Split install-sidebar-banner into install-perk and install-card
- Drop the "browser not supported" banner in favour of the manual instructions
- Hide the install perk from anyone already running the installed app

beforeinstallprompt is Chromium-only, so treating its absence as "unsupported"
was wrong for macOS Safari, which installs through File then Add to Dock, and for
Firefox on Android, which installs from its own menu. Both were told to switch
browser. Everything without a native prompt now falls through to
InstallInstructionsSheet, whose non-iOS branch already spelled out those steps and
was simply unreachable from this path.

The perk branch never checked isStandalone, unlike the card, so someone already
inside the installed app was still invited to install it.

One file branched on two orthogonal props into four renders, and the matrix had a
corner that reached none of them. Two components make that combination
unrepresentable.
